added hoist functions and fix included files
