--- babel.config.js
@@ -1,5 +1,0 @@
-const presets = [
-  'babel/preset-react'
-];
-
-module.exports = { presets };
