+++ package.json
@@ -6,9 +6,11 @@
   "scripts": {
     "test": "jest",
     "lint": "eslint src --fix ",
-    "build": "npm run lint && jest && npm run babel",
     "minorUpdate": "npm run build && npm version minor && npm publish"
   },
+  "files": [
+    "src"
+  ],
   "repository": {
     "type": "git",
     "url": "https://github.com/Niryo/controllerim.git"
@@ -20,6 +22,7 @@
   },
   "dependencies": {
     "babel-runtime": "^6.26.0",
+    "hoist-non-react-statics": "^3.3.1",
     "lodash": "^4.17.10",
     "mobx": "^3.6.2"
   },
