+++ src/index.js
@@ -1,4 +1,4 @@
 export { observer } from './Observer/observer';
 export { Controller } from './Controller/Controller';
-export { TestUtils } from './TestUtils/testUtils';
+export { TestUtils } from '../test/TestUtils/testUtils';
 export { useExperimentalSerialization } from './AutoIndexManager/AutoIndexManager';
